refactor: エントリー取得処理や型を別ファイルに移動
