+++ src/db/queries/journals.ts
@@ -16,16 +16,6 @@ export const getJournalsQuery = db.query.journals.findMany({
 });
 
 /**
- * ジャーナルに紐付いたエントリー一覧を取得するクエリ
- * @param journalId ジャーナルID
- */
-export const getEntriesQuery = (journalId: string) =>
-  db.query.journals.findFirst({
-    where: (journals, { eq }) => eq(journals.id, journalId),
-    with: { entries: true },
-  });
-
-/**
  * ジャーナルをフィールドと共に作成するクエリを実行
  * @param journal ジャーナルのメタ情報
  * @param newFields フィールド一覧
