+++ src/utils/journal/use-entry.ts
@@ -1,0 +1,7 @@
+export type EntryObj = {
+  id: string;
+  date: Date;
+  title: string;
+  preview: string;
+  bookmark?: boolean;
+};
